refactor(counters): move form helpers to counter-form module
Lift UserError, normalizeName, namedValue, bucketTarget, bucketLabel and
resolveAddTarget out of the route into lib/server/counter-form.ts, so
+page.server.ts is a thin list of actions again and drops back under the
file overall-complexity gate. No behaviour change.
